rm redux and redux-saga
