+++ apps/learneth/src/App.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, {useEffect, useReducer} from 'react'
 import {createHashRouter, RouterProvider} from 'react-router-dom'
 import {ToastContainer} from 'react-toastify'
 import LoadingScreen from './components/LoadingScreen'
@@ -6,6 +6,9 @@ import LogoPage from './pages/Logo'
 import HomePage from './pages/Home'
 import StepListPage from './pages/StepList'
 import StepDetailPage from './pages/StepDetail'
+import {appInitialState, appReducer} from './reducers/state'
+import {updateState, initDispatch, connectRemix} from './actions'
+import {AppContext} from './contexts'
 import 'react-toastify/dist/ReactToastify.css'
 import './App.css'
 
@@ -29,12 +32,26 @@ export const router = createHashRouter([
 ])
 
 function App(): JSX.Element {
+  const [appState, dispatch] = useReducer(appReducer, appInitialState)
+  useEffect(() => {
+    updateState(appState)
+  }, [appState])
+  useEffect(() => {
+    initDispatch(dispatch)
+    updateState(appState)
+    connectRemix()
+  }, [])
   return (
-    <>
+    <AppContext.Provider
+      value={{
+        dispatch,
+        appState,
+      }}
+    >
       <RouterProvider router={router} />
       <LoadingScreen />
       <ToastContainer position="bottom-right" newestOnTop closeOnClick rtl={false} pauseOnFocusLoss draggable pauseOnHover autoClose={false} theme="colored" />
-    </>
+    </AppContext.Provider>
   )
 }
