+++ apps/learneth/src/contexts/index.ts
@@ -1,0 +1,3 @@
+import {createContext} from 'react'
+
+export const AppContext = createContext<any>({})
