+++ apps/learneth/src/remix-client.ts
@@ -1,6 +1,6 @@
 import {PluginClient} from '@remixproject/plugin'
 import {createClient} from '@remixproject/plugin-webview'
-import {store} from './redux/store'
+import {loadRepo} from './actions'
 import {router} from './App'
 
 class RemixClient extends PluginClient {
@@ -12,25 +12,19 @@ class RemixClient extends PluginClient {
   startTutorial(name: any, branch: any, id: any): void {
     console.log('start tutorial', name, branch, id)
     void router.navigate('/home')
-    store.dispatch({
-      type: 'workshop/loadRepo',
-      payload: {
-        name,
-        branch,
-        id,
-      },
+    loadRepo({
+      name,
+      branch,
+      id,
     })
   }
 
   addRepository(name: any, branch: any) {
     console.log('add repo', name, branch)
     void router.navigate('/home')
-    store.dispatch({
-      type: 'workshop/loadRepo',
-      payload: {
-        name,
-        branch,
-      },
+    loadRepo({
+      name,
+      branch,
     })
   }
 }
