+++ package.json
@@ -103,7 +103,6 @@
     "@openzeppelin/contracts": "^5.0.0",
     "@openzeppelin/upgrades-core": "^1.30.0",
     "@openzeppelin/wizard": "0.4.0",
-    "@reduxjs/toolkit": "^2.0.1",
     "@remixproject/engine": "0.3.42",
     "@remixproject/engine-electron": "0.3.42",
     "@remixproject/engine-web": "0.3.42",
@@ -184,7 +183,6 @@
     "react-virtuoso": "^4.6.2",
     "react-window": "^1.8.10",
     "react-zoom-pan-pinch": "^3.1.0",
-    "redux-saga": "^1.3.0",
     "regenerator-runtime": "0.13.7",
     "rehype-raw": "^6.0.0",
     "remark-gfm": "^3.0.1",
